Optimize non-streaming XML parser
